refactor(VDatePickerMonth): integrate calendar composable
